fix: expose more config options to ngxway.conf

+++ example/goserver/build.sh
@@ -1,7 +1,10 @@
-wget https://dl.google.com/go/go1.15.linux-amd64.tar.gz
-tar -C /usr/local -xzf go1.15.linux-amd64.tar.gz
+# Check go is installed.
+if ! command -v go &> /dev/null; then
+  wget https://dl.google.com/go/go1.15.linux-amd64.tar.gz
+  tar -C /usr/local -xzf go1.15.linux-amd64.tar.gz
+  rm -f go1.15.linux-amd64.tar.gz
+fi
+
 export PATH=$PATH:/usr/local/go/bin
 go env -w GO111MODULE=on
-rm -f go1.15.linux-amd64.tar.gz
-
 go build -o goserver main.go
